Image picker intent

+++ PatientClient/app/src/main/java/com/team13/patientclient/models/PatientModel.java
@@ -69,7 +69,7 @@ public class PatientModel {
         this.gender = gender;
     }
 
-    public double getDateOfBirth() {
+    public long getDateOfBirth() {
         return dateOfBirth;
     }
